feat core: validate error page headers at service start
`HttpResponse::SetHeader` rejects header names and values with characters
that RFC 9110 does not allow, and the error page headers come straight
from the static config. A typo there used to surface only when a request
actually hit that page, throwing out of the failsafe task on every such
request instead of once at start.

Move the two checks out of the anonymous namespace of http_response.cpp
into header_validation.hpp so the config parsing can reuse them, and run
them on every configured error page header.
